+++ crates/cli/src/lib.rs
@@ -54,10 +54,6 @@ pub use {args::Args, error::Error, value::Value};
 /// Crate-specific result type.
 pub type Result<T> = std::result::Result<T, Error>;
 
-#[cfg(feature = "duckdb")]
-use duckdb as _;
-use openssl as _;
-use openssl_src as _;
 use tracing_subscriber as _;
 
 #[cfg(test)]