+++ crates/pgstac/src/lib.rs
@@ -76,15 +76,11 @@
 #![warn(missing_docs)]
 
 mod page;
-#[cfg(feature = "tls")]
-mod tls;
 
 pub use page::Page;
 use serde::{de::DeserializeOwned, Serialize};
 use stac_api::Search;
 use tokio_postgres::{types::ToSql, GenericClient, Row};
-#[cfg(feature = "tls")]
-pub use {tls::make_unverified_tls, tokio_postgres_rustls::MakeRustlsConnect};
 
 /// Crate-specific error enum.
 #[derive(Debug, thiserror::Error)]
@@ -171,6 +167,11 @@ pub trait Pgstac: GenericClient {
     {
         let collection = serde_json::to_value(collection)?;
         self.pgstac_void("upsert_collection", &[&collection]).await
+    }
+
+    /// Updates all collection extents.
+    async fn update_collection_extents(&self) -> Result<()> {
+        self.pgstac_void("update_collection_extents", &[]).await
     }
 
     /// Updates a collection.
@@ -351,13 +352,17 @@ pub(crate) mod tests {
         dbname: String,
     }
 
+    pub(crate) fn config() -> Config {
+        std::env::var("PGSTAC_RS_TEST_DB")
+            .unwrap_or("postgresql://username:password@localhost:5432/postgis".to_string())
+            .parse()
+            .unwrap()
+    }
+
     impl TestClient {
         async fn new(id: u16) -> TestClient {
             let dbname = format!("pgstac_test_{}", id);
-            let config: Config = std::env::var("PGSTAC_RS_TEST_DB")
-                .unwrap_or("postgresql://username:password@localhost:5432/postgis".to_string())
-                .parse()
-                .unwrap();
+            let config = config();
             {
                 let _mutex = MUTEX.lock().unwrap();
                 let (client, connection) = config.connect(NoTls).await.unwrap();
@@ -571,6 +576,7 @@ pub(crate) mod tests {
                 .unwrap(),
             serde_json::to_value(item).unwrap(),
         );
+        client.update_collection_extents().await.unwrap();
     }
 
     #[rstest]
@@ -704,12 +710,12 @@ pub(crate) mod tests {
         item.geometry = Some(longmont());
         client.add_item(item.clone()).await.unwrap();
         let search = Search {
-            ids: Some(vec!["an-id".to_string()]),
+            ids: vec!["an-id".to_string()],
             ..Default::default()
         };
         assert_eq!(client.search(search).await.unwrap().features.len(), 1);
         let search = Search {
-            ids: Some(vec!["not-an-id".to_string()]),
+            ids: vec!["not-an-id".to_string()],
             ..Default::default()
         };
         assert!(client.search(search).await.unwrap().features.is_empty());
@@ -725,12 +731,12 @@ pub(crate) mod tests {
         item.geometry = Some(longmont());
         client.add_item(item.clone()).await.unwrap();
         let search = Search {
-            collections: Some(vec!["collection-id".to_string()]),
+            collections: vec!["collection-id".to_string()],
             ..Default::default()
         };
         assert_eq!(client.search(search).await.unwrap().features.len(), 1);
         let search = Search {
-            collections: Some(vec!["not-an-id".to_string()]),
+            collections: vec!["not-an-id".to_string()],
             ..Default::default()
         };
         assert!(client.search(search).await.unwrap().features.is_empty());
@@ -913,12 +919,12 @@ pub(crate) mod tests {
         item.id = "b".to_string();
         client.add_item(item).await.unwrap();
         let mut search = Search::default();
-        search.items.sortby = Some(vec![Sortby::asc("id")]);
+        search.items.sortby = vec![Sortby::asc("id")];
         let page = client.search(search.clone()).await.unwrap();
         assert_eq!(page.features[0]["id"], "a");
         assert_eq!(page.features[1]["id"], "b");
 
-        search.items.sortby = Some(vec![Sortby::desc("id")]);
+        search.items.sortby = vec![Sortby::desc("id")];
         let page = client.search(search).await.unwrap();
         assert_eq!(page.features[0]["id"], "b");
         assert_eq!(page.features[1]["id"], "a");